# specs/001-photo-album-organizer/contracts/api.yaml
openapi: 3.0.3
info:
  title: Photo Album Organizer API
  description: Backend API for photo album management and retrieval
  version: 1.0.0
  contact:
    name: Photo Album Organizer Team
servers:
  - url: http://localhost:3001/api
    description: Local development server
paths:
...
  /albums/{id}:
    get:
      summary: Get album by ID
      description: Returns album details with all photos
      operationId: getAlbumById
      tags:
        - Albums
      parameters:
        - name: id
          in: path
          required: true
          description: Album ID
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlbumDetail'
              example:
                id: 1
                city_name: "Paris"
                photos:
                  - id: 101
                    file_path: "/sample-photos/paris-1.jpg"
                    display_order: 0
                  - id: 102
                    file_path: "/sample-photos/paris-2.jpg"
                    display_order: 1
        '404':
          description: Album not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Album not found"
                code: "ALBUM_NOT_FOUND"
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    AlbumDetail:
      type: object
      description: Album with all photos
      required:
        - id
        - city_name
        - photos
      properties:
        id:
          type: integer
          description: Unique album identifier
          example: 1
        city_name:
          type: string
          description: Name of the city
          minLength: 1
          maxLength: 50
          example: "Paris"
        photos:
          type: array
          description: Array of photos in display order
          items:
            $ref: '#/components/schemas/Photo'
    Photo:
      type: object
      description: Photo metadata
      required:
        - id
        - file_path
        - display_order
      properties:
        id:
          type: integer
          description: Unique photo identifier
          example: 101
        file_path:
          type: string
          description: Relative path to photo file
          example: "/sample-photos/paris-1.jpg"
        display_order:
          type: integer
          description: Order within album (0-based)
          minimum: 0
          example: 0
...
    Error:
      type: object
      description: Error response
      required:
        - error
      properties:
        error:
          type: string
          description: Human-readable error message
          example: "Album not found"
        code:
          type: string
          description: Machine-readable error code
          example: "ALBUM_NOT_FOUND"
        details:
          type: object
          description: Additional error context
          nullable: true
...